Fix directory exclusions when check existence
Checking if a file name 'is a file', the directories were being
excluded, when they are part of the scope. I use instead the 'exist'
check of the 'os.path' library.

To do the code concordant with the generality of the purpose (file and
directories and not only files) I also change the variable names from
'filename' to 'path'.

This commit fixes issue #26.